feat: implement simplified progress bar with better default layout
The columns have been added to make it closer to what tqdm uses, which
make so much more sense than rich's default. For example, the ETA was
only calculated based on the last step and if it exceeded 30s it would
always be empty.

Also the API of rich is really annoying for a simple progress bar, so
this fixes that.