fix(manager): remove duplicate renders when re-enabling the manager
When the manager was done with the given tasks, it kept the widgets that
just finished, so when it was re-activated, it would be rendered again,
creating duplicate renderings.
Therefore, once the manager is disabled, the persisting widgets are
rendered and after the live was closed, they will be removed from the
manager.